Initial work on hopshackle1
We have a simple featureSet that consists of a 1-mesh in the 8 squares surrounding the Avatar, and a 3-mesh covering (with no overlap) the 9x9 area centered on the Avatar.
We now use this to construct trajectories, and use a policy to choose actions.
However we do not yet do any learning of a policy.